Support alpha minimally
This is effectively what perceptualdiff used to do before FreeImage was
introduced. Years ago before this happened, perceptualdiff would store
color with premultiplied alphas.

I call this support minimal since it won't detect differences in alpha
where the pixel color is black. This is because the alpha multiplied by
a zero color will result in zero no matter what the alpha is.

This fixes #12.